requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-Session==0.5.0
PyMySQL==1.1.0
cryptography==41.0.7
Werkzeug==3.0.1
python-dotenv==1.0.0
Flask-Mail==0.9.1
